adding object picking, something is off but it's close?

+++ Engine/Source/Util/UUID.cpp
@@ -7,20 +7,20 @@ namespace Cosmos
 {
 	static std::random_device sDevice;
 	static std::mt19937_64 sEngine(sDevice());
-	static std::uniform_int_distribution<uint64_t> sDistribution;
+	static std::uniform_int_distribution<uint32_t> sDistribution;
 
 	UUID::UUID()
 		: mUUID(sDistribution(sEngine))
 	{
 	}
 
-	UUID::UUID(uint64_t id)
+	UUID::UUID(uint32_t id)
 		: mUUID(id)
 	{
 	}
 
 	UUID::UUID(std::string id)
 	{
-		mUUID = std::stoull(id);
+		mUUID = std::stoi(id);
 	}
 }
